Remove code generation from casm.go
The Casm structure is split and the code generation is put his specific file. A copper code generator and x86-64 code generator are added, so we can build for different targets.
