+++ pkg/casm/copper_generator_test.go
@@ -7,125 +7,8 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestSaveProgramToFile(t *testing.T) {
-	tests := []struct {
-		casm     Casm
-		hasError bool
-	}{
-		{Casm{OutputFile: "testdata/test.notcopper"}, true},
-		{Casm{OutputFile: "testdata/test.copper"}, false},
-	}
-
-	for _, test := range tests {
-		err := test.casm.SaveProgramToFile()
-
-		if test.hasError {
-			assert.Error(t, err, test)
-		} else {
-			assert.NoError(t, err, test)
-		}
-	}
-}
-
-func TestTranslateSourceFile(t *testing.T) {
-	tests := []struct {
-		path     string
-		hasError bool
-	}{
-		{"testdata/test.notcasm", true},
-		{"testdata/test1.casm", true},
-		{"testdata/test.casm", false},
-	}
-	casm := Casm{}
-
-	for _, test := range tests {
-		err := casm.TranslateSourceFile(test.path)
-
-		if test.hasError {
-			assert.Error(t, err, test)
-		} else {
-			assert.NoError(t, err, test)
-		}
-	}
-}
-
-func ir(kind IRKind, value interface{}, loc FileLocation) (ir IR) {
-	ir.Location = loc
-	ir.Kind = kind
-	switch kind {
-	case IRKindLabel:
-		ir.AsLabel = value.(LabelIR)
-	case IRKindInstruction:
-		ir.AsInstruction = value.(InstructionIR)
-	case IRKindEntry:
-		ir.AsEntry = value.(EntryIR)
-	case IRKindConst:
-		ir.AsConst = value.(ConstIR)
-	case IRKindMemory:
-		ir.AsMemory = value.(MemoryIR)
-	}
-	return ir
-}
-
-var testSources = []struct {
-	in       string
-	out      []IR
-	hasError bool
-}{
-	{"main:\n", []IR{ir(IRKindLabel, LabelIR{"main"}, FileLocation{"test_file", 1})}, false},
-	{"push 1\n", []IR{
-		ir(IRKindInstruction, InstructionIR{
-			Name:       "push",
-			HasOperand: true,
-			Operand:    expression(ExpressionKindNumLitInt, int64(1)),
-		}, FileLocation{"test_file", 1}),
-	}, false},
-	{"%const N 1\n", []IR{ir(IRKindConst, ConstIR{
-		"N",
-		expression(ExpressionKindNumLitInt, int64(1)),
-	}, FileLocation{"test_file", 1})}, false},
-	{":", []IR{}, true},
-	{"wrong\n", []IR{}, true},
-	{"push \n", []IR{}, true},
-	{"%dir 0\n", []IR{}, true},
-	{"push N\n", []IR{
-		ir(IRKindInstruction, InstructionIR{
-			HasOperand: true,
-			Name:       "push",
-			Operand:    expression(ExpressionKindBinding, "N"),
-		}, FileLocation{"test_file", 1}),
-	}, false},
-	{"%include abc", []IR{}, true},
-}
-
-func TestTranslateIR(t *testing.T) {
-	for _, test := range testSources {
-		func() {
-			defer func() {
-				r := recover()
-				if r != nil && !test.hasError {
-					assert.Fail(t, "unexpected error", test)
-				}
-			}()
-
-			lines, err := Linize(test.in, "test_file")
-			if err != nil {
-				panic(err)
-			}
-
-			ctx := Casm{}
-			irs := ctx.translateIR(lines)
-
-			if test.hasError {
-				assert.Fail(t, "expecting an error", test)
-			}
-			assert.Equal(t, test.out, irs, test)
-		}()
-	}
-}
-
 func TestGetBindingByName(t *testing.T) {
-	casm := Casm{
+	cgen := copperGenerator{
 		Bindings: []Binding{
 			{Name: "a_label",
 				Value:    expression(ExpressionKindNumLitInt, int64(1)),
@@ -138,20 +21,20 @@ func TestGetBindingByName(t *testing.T) {
 		},
 	}
 
-	exist, binding := casm.getBindingByName("a_label")
+	exist, binding := cgen.getBindingByName("a_label")
 	assert.True(t, exist)
-	assert.Equal(t, casm.Bindings[0], binding)
+	assert.Equal(t, cgen.Bindings[0], binding)
 
-	exist, binding = casm.getBindingByName("a_const")
+	exist, binding = cgen.getBindingByName("a_const")
 	assert.True(t, exist)
-	assert.Equal(t, casm.Bindings[1], binding)
+	assert.Equal(t, cgen.Bindings[1], binding)
 
-	exist, binding = casm.getBindingByName("test")
+	exist, binding = cgen.getBindingByName("test")
 	assert.False(t, exist)
 }
 
 func TestGetBindingIndexByName(t *testing.T) {
-	casm := Casm{
+	cgen := copperGenerator{
 		Bindings: []Binding{
 			{Name: "a_label",
 				Value:    expression(ExpressionKindNumLitInt, int64(1)),
@@ -164,18 +47,18 @@ func TestGetBindingIndexByName(t *testing.T) {
 		},
 	}
 
-	index := casm.getBindingIndexByName("a_label")
+	index := cgen.getBindingIndexByName("a_label")
 	assert.Equal(t, 0, index)
 
-	index = casm.getBindingIndexByName("a_const")
+	index = cgen.getBindingIndexByName("a_const")
 	assert.Equal(t, 1, index)
 
-	index = casm.getBindingIndexByName("test")
+	index = cgen.getBindingIndexByName("test")
 	assert.Equal(t, -1, index)
 }
 
 func TestBindLabel(t *testing.T) {
-	casm := Casm{
+	cgen := copperGenerator{
 		Bindings: []Binding{
 			{Name: "a_label",
 				Value:    expression(ExpressionKindNumLitInt, int64(1)),
@@ -186,12 +69,12 @@ func TestBindLabel(t *testing.T) {
 
 	func() {
 		defer func() { recover() }()
-		casm.bindLabel(LabelIR{"a_label"}, 1, FileLocation{})
+		cgen.bindLabel(LabelIR{"a_label"}, 1, FileLocation{})
 		assert.Fail(t, "expecting an error")
 	}()
 
-	casm.bindLabel(LabelIR{"new_label"}, 2, FileLocation{})
-	exist, binding := casm.getBindingByName("new_label")
+	cgen.bindLabel(LabelIR{"new_label"}, 2, FileLocation{})
+	exist, binding := cgen.getBindingByName("new_label")
 	assert.True(t, exist)
 
 	label := Binding{
@@ -238,7 +121,7 @@ func TestBindConst(t *testing.T) {
 				}
 			}()
 
-			casm := Casm{
+			cgen := copperGenerator{
 				Bindings: []Binding{
 					{Name: "a_const",
 						Value:    Expression{Kind: ExpressionKindNumLitInt, AsNumLitInt: 1},
@@ -246,15 +129,15 @@ func TestBindConst(t *testing.T) {
 						IsLabel:  false},
 				},
 			}
-			casm.bindConst(test.constIR, FileLocation{})
+			cgen.bindConst(test.constIR, FileLocation{})
 
 			if test.hasError {
 				assert.Fail(t, "expecting an error", test)
 			} else {
-				exist, binding := casm.getBindingByName(test.constIR.Name)
+				exist, binding := cgen.getBindingByName(test.constIR.Name)
 				assert.True(t, exist, test)
 				assert.Equal(t, test.binding, binding, test)
-				assert.Equal(t, test.memoryLength, len(casm.Memory), test)
+				assert.Equal(t, test.memoryLength, len(cgen.Memory), test)
 			}
 		}()
 	}
@@ -262,22 +145,22 @@ func TestBindConst(t *testing.T) {
 
 func TestBindEntry(t *testing.T) {
 	func() {
-		casm := Casm{}
+		cgen := copperGenerator{}
 		defer func() { recover() }()
-		casm.HasEntry = true
-		casm.bindEntry(EntryIR{"main"}, FileLocation{})
+		cgen.HasEntry = true
+		cgen.bindEntry(EntryIR{"main"}, FileLocation{})
 		assert.Fail(t, "expecting an error")
 	}()
 
-	casm := Casm{}
-	casm.bindEntry(EntryIR{"entry"}, FileLocation{"", 10})
-	assert.True(t, casm.HasEntry)
-	assert.Equal(t, "entry", casm.DeferredEntryName)
-	assert.Equal(t, 10, casm.EntryLocation.Location)
+	cgen := copperGenerator{}
+	cgen.bindEntry(EntryIR{"entry"}, FileLocation{"", 10})
+	assert.True(t, cgen.HasEntry)
+	assert.Equal(t, "entry", cgen.DeferredEntryName)
+	assert.Equal(t, 10, cgen.EntryLocation.Location)
 }
 
 func TestBindMemory(t *testing.T) {
-	casm := Casm{
+	cgen := copperGenerator{
 		Bindings: []Binding{
 			{Name: "mem",
 				Value:    expression(ExpressionKindNumLitInt, int64(0)),
@@ -288,12 +171,12 @@ func TestBindMemory(t *testing.T) {
 
 	func() {
 		defer func() { recover() }()
-		casm.bindMemory(MemoryIR{Name: "mem", Value: expression(ExpressionKindByteList, []byte{1})}, FileLocation{})
+		cgen.bindMemory(MemoryIR{Name: "mem", Value: expression(ExpressionKindByteList, []byte{1})}, FileLocation{})
 		assert.Fail(t, "expecting an error")
 	}()
 
-	casm.bindMemory(MemoryIR{Name: "new_mem", Value: expression(ExpressionKindByteList, []byte{2, 3})}, FileLocation{})
-	exist, binding := casm.getBindingByName("new_mem")
+	cgen.bindMemory(MemoryIR{Name: "new_mem", Value: expression(ExpressionKindByteList, []byte{2, 3})}, FileLocation{})
+	exist, binding := cgen.getBindingByName("new_mem")
 	assert.True(t, exist)
 
 	want := Binding{
@@ -320,8 +203,8 @@ var evaluateExpressionsTests = []struct {
 }
 
 func TestEvaluateExpression(t *testing.T) {
-	casm := Casm{}
-	casm.Bindings = append(casm.Bindings, Binding{
+	cgen := copperGenerator{}
+	cgen.Bindings = append(cgen.Bindings, Binding{
 		Name:  "a_bind",
 		Value: expression(ExpressionKindNumLitInt, int64(3)),
 	})
@@ -335,7 +218,7 @@ func TestEvaluateExpression(t *testing.T) {
 				}
 			}()
 
-			result := casm.evaluateExpression(test.expr, FileLocation{}).Word
+			result := cgen.evaluateExpression(test.expr, FileLocation{}).Word
 
 			if test.hasError {
 				assert.Fail(t, "expecting an error", test)
@@ -579,7 +462,7 @@ var binaryOpTests = []struct {
 
 func TestEvaluateBinaryOp(t *testing.T) {
 	for _, test := range binaryOpTests {
-		casm := Casm{}
+		cgen := copperGenerator{}
 		func() {
 			defer func() {
 				r := recover()
@@ -588,7 +471,7 @@ func TestEvaluateBinaryOp(t *testing.T) {
 				}
 			}()
 
-			word := casm.evaluateBinaryOp(test.expr, FileLocation{}).Word
+			word := cgen.evaluateBinaryOp(test.expr, FileLocation{}).Word
 
 			if test.hasError {
 				assert.Fail(t, "expecting an error", test)
@@ -600,7 +483,7 @@ func TestEvaluateBinaryOp(t *testing.T) {
 }
 
 func TestEvaluateBinding(t *testing.T) {
-	casm := Casm{
+	cgen := copperGenerator{
 		Bindings: []Binding{
 			{Name: "a_bind", Value: expression(ExpressionKindNumLitInt, int64(5))},
 			{Name: "cycl1", Value: expression(ExpressionKindBinding, "cycl2")},
@@ -608,55 +491,45 @@ func TestEvaluateBinding(t *testing.T) {
 		},
 	}
 
-	word := casm.evaluateBinding(casm.Bindings[0], FileLocation{}).Word
+	word := cgen.evaluateBinding(cgen.Bindings[0], FileLocation{}).Word
 	assert.Equal(t, coppervm.WordU64(5), word)
 
-	word = casm.evaluateBinding(casm.Bindings[0], FileLocation{}).Word
+	word = cgen.evaluateBinding(cgen.Bindings[0], FileLocation{}).Word
 	assert.Equal(t, coppervm.WordU64(5), word)
 
 	func() {
 		defer func() { recover() }()
-		casm.evaluateBinding(casm.Bindings[1], FileLocation{})
+		cgen.evaluateBinding(cgen.Bindings[1], FileLocation{})
 		assert.Fail(t, "expecting an error")
 	}()
 
 	func() {
 		defer func() { recover() }()
-		casm.evaluateBinding(Binding{Name: "bind"}, FileLocation{})
+		cgen.evaluateBinding(Binding{Name: "bind"}, FileLocation{})
 		assert.Fail(t, "expecting an error")
 	}()
-}
-
-func TestStrings(t *testing.T) {
-	casm := Casm{}
-	casm.TranslateSourceFile("testdata/string.casm")
-	assert.Equal(t, uint64(0), casm.Program[0].Operand.AsU64)
-	assert.Equal(t, uint64(0), casm.Program[1].Operand.AsU64)
-	assert.Equal(t, uint64(9), casm.Program[2].Operand.AsU64)
-	assert.Equal(t, uint64(24), casm.Program[3].Operand.AsU64)
-	assert.Equal(t, 34, len(casm.Memory))
 }
 
 func TestPushStringToMemory(t *testing.T) {
-	casm := Casm{}
+	cgen := copperGenerator{}
 
-	strAddr := casm.pushStringToMemory("string1")
+	strAddr := cgen.pushStringToMemory("string1")
 	assert.Equal(t, 0, strAddr)
-	assert.Equal(t, 8, casm.StringLengths[0])
+	assert.Equal(t, 8, cgen.StringLengths[0])
 
-	strAddr = casm.pushStringToMemory("a different string")
+	strAddr = cgen.pushStringToMemory("a different string")
 	assert.Equal(t, 8, strAddr)
-	assert.Equal(t, 19, casm.StringLengths[8])
+	assert.Equal(t, 19, cgen.StringLengths[8])
 }
 
 func TestGetStringByAddress(t *testing.T) {
-	casm := Casm{}
-	l1 := casm.pushStringToMemory("string1")
-	s1 := casm.getStringByAddress(l1)
+	cgen := copperGenerator{}
+	l1 := cgen.pushStringToMemory("string1")
+	s1 := cgen.getStringByAddress(l1)
 	assert.NotEmpty(t, s1)
 	assert.Equal(t, "string1", s1)
-	assert.Len(t, casm.Memory, 8)
+	assert.Len(t, cgen.Memory, 8)
 
-	s2 := casm.getStringByAddress(1)
+	s2 := cgen.getStringByAddress(1)
 	assert.Empty(t, s2)
 }
