+++ pkg/casm/target.go
@@ -1,0 +1,8 @@
+package casm
+
+type BuildTarget int
+
+const (
+	BuildTargetCopper BuildTarget = iota
+	BuildTargetX86_64
+)
