--- pkg/casm/testdata/string.casm
@@ -1,5 +1,0 @@
-%const str "a string"
-push str
-push str
-push "a new string"
-push "a"+str
